pyproject.toml
[project]
name = "crash-mcp"
version = "0.2.0"
description = "MCP server for crash dump analysis"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"

dependencies = [
    "mcp",
    "python-dotenv",
    "pexpect",
    "drgn",
    "drgn-tools",
    "click",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
crash-mcp = "crash_mcp.server:main"
compile-crash = "crash_mcp.builder.compile_crash:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
  "src/crash_mcp/resource/*",
]

[tool.hatch.build.targets.wheel]
packages = ["src/crash_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

